+++ src/main/java/com/zx/sms/codec/smpp/msg/Pdu.java
@@ -57,7 +57,7 @@ public abstract class Pdu implements BaseMessage,Cloneable{
         this.isRequest = isRequest;
         this.commandLength = null;
         this.commandId = commandId;
-        this.sequenceNumber = (int)DefaultSequenceNumberUtil.getSequenceNo();
+        this.sequenceNumber = DefaultSequenceNumberUtil.getSequenceNo();
         this.referenceObject = null;
     }
 
@@ -383,11 +383,11 @@ public abstract class Pdu implements BaseMessage,Cloneable{
     }
     
 	@Override
-	public long getSequenceNo() {
+	public int getSequenceNo() {
 		return getSequenceNumber();
 	}
 	
-	public void setSequenceNo(long seq) {
-		setSequenceNumber((int)seq);
+	public void setSequenceNo(int seq) {
+		setSequenceNumber(seq);
 	}
 }
